MSEARCH-418: Support interface instance-storage 9.0
mod-inventory-storage has new major interface versions:

instance-storage 9.0
holdings-storage 6.0
item-storage 10.0

mod-search uses only instance-storage, not the other two.

mod-search is not affected because the incompatible change is in the DELETE all APIs only that is not used by mod-search.

Only the okapiInterfaces need to be bumped in "requires" section of ModuleDescriptor-template.json