MSEARCH-415: Option for Elasticsearch integration test
In #cc-elasticsearch-opensearch-discussion Slack channel people asked whether mod-search can still run with an Elasticsearch server.

Provide an option to switch the search engine from OpenSearch to Elasticsearch when running the integration test phase of the build.

Note that this pull request uses old Elasticsearch 8.0.1 container because the current 8.3.3 container has a change:
Since version 8.1 Elasticsearch no longer generates an empty array but an undefined property making many
unit tests fail. Upgrading Elasticseach from 8.0.1 to 8.3.3 will be handled in a separate issue and PR.

This obsoletes PR #289 that tried to combine the option and the upgrade from 8.0.1 to 8.3.3.