MSEARCH-427: Add integration test option for Elasticsearch 8
## Purpose

Elasticsearch 7.17.* is LTS with EOL date 2023-08-01: https://www.elastic.co/support/eol#elasticsearch

However, people on #cc-elasticsearch-opensearch-discussion Slack channel asked for the compatibility with latest Elasticsearch version.

## Approach

Add an additional Dockerfile for the latest Elasticsearch 8 version (8.4.1) one can manually switch to before running the mod-search integration tests.
The README.md is updated to mention this new option.

OpenSearch (all versions) and Elasticsearch before 8.1.0 returns an empty list when a list join yields no element.
Elasticsearch from 8.1.0 on returns null when a list join yields no element.

This fails a few tests:
* BrowseCallNumberIT.browseByCalNumber_browseAroundWithEnabledIntermediateValues:111
* BrowseCallNumberIT.browseByCallNumber_browsingAroundWhenPrecedingRecordsCountIsSpecified:72
* BrowseCallNumberIT.browseByCallNumber_browsingAroundWithoutHighlightMatch:91
* BrowseCallNumberIT.browseByCallNumber_parameterized:61
* BrowseCallNumberOtherIT.browseByCallNumber_browsingAroundWithDisabledIntermediateValues:53
* BrowseCallNumberOtherIT.browseByCallNumber_browsingAroundWithEnabledIntermediateValues:78
* SearchInstanceIT.responseContainsAllInstanceProperties:136

To compensate for this change in Elasticsearch 8.1.0 mod-search's API schemas are changed:
The affected list properties are declared as required, this automatically changes a null
result into an empty list.

The returned JSON is the same as before.

The schemas are used for GET APIs only, therefore this schema change is a non-breaking change.
And the JSON response is the same as before.

Quote from https://www.elastic.co/guide/en/elasticsearch/reference/current/null-value.html
"A null value cannot be indexed or searched. When a field is set to null, (or an empty array
or an array of null values) it is treated as though that field has no values."

### Changes checklist

- [ ] API paths, methods, request or response bodies changed, added, or removed
- [ ] Database schema changes
- [ ] Interface versions changes
- [ ] Interface dependencies added, or removed
- [ ] Permissions changed, added, or removed